+++ server/src/main/java/no/fractal/socket/FractalClient.java
@@ -1,11 +1,11 @@
 package no.fractal.socket;
 
 import no.fractal.socket.meta.*;
-import no.fractal.socket.payload.AuthenticationPayload;
+import no.fractal.socket.messages.recive.AuthenticationPayload;
 import no.fractal.socket.payload.InvalidPayloadException;
-import no.fractal.socket.payload.NoSuchPayloadException;
-import no.fractal.socket.payload.PayloadBase;
-import no.fractal.socket.payload.ThumbnailPayload;
+import no.fractal.socket.messages.recive.NoSuchPayloadException;
+import no.fractal.socket.messages.recive.PayloadBase;
+import no.fractal.socket.messages.recive.ThumbnailPayload;
 
 import java.io.BufferedInputStream;
